refactoring: make `ResourceList` type reusable by both primary ADS server and future MADS server
