+++ pkg/xds/generator/inbound_proxy_generator_test.go
@@ -99,7 +99,7 @@ var _ = Describe("InboundProxyGenerator", func() {
 			Expect(err).ToNot(HaveOccurred())
 
 			// then
-			resp := generator.ResourceList(rs).ToDeltaDiscoveryResponse()
+			resp := model.ResourceList(rs).ToDeltaDiscoveryResponse()
 			actual, err := util_proto.ToYAML(resp)
 			Expect(err).ToNot(HaveOccurred())
