+++ pkg/xds/generator/outbound_proxy_generator_test.go
@@ -119,7 +119,7 @@ var _ = Describe("OutboundProxyGenerator", func() {
 			Expect(err).ToNot(HaveOccurred())
 
 			// then
-			resp := generator.ResourceList(rs).ToDeltaDiscoveryResponse()
+			resp := model.ResourceList(rs).ToDeltaDiscoveryResponse()
 			actual, err := util_proto.ToYAML(resp)
 			Expect(err).ToNot(HaveOccurred())
