+++ pkg/xds/generator/prometheus_endpoint_generator_test.go
@@ -140,7 +140,7 @@ var _ = Describe("PrometheusEndpointGenerator", func() {
 			Expect(err).ToNot(HaveOccurred())
 
 			// then
-			resp := generator.ResourceList(rs).ToDeltaDiscoveryResponse()
+			resp := model.ResourceList(rs).ToDeltaDiscoveryResponse()
 			actual, err := util_proto.ToYAML(resp)
 
 			Expect(err).ToNot(HaveOccurred())
