+++ pkg/xds/generator/proxy_template_raw_source_test.go
@@ -210,7 +210,7 @@ var _ = Describe("ProxyTemplateRawSource", func() {
 			Expect(err).ToNot(HaveOccurred())
 
 			// then
-			resp := generator.ResourceList(rs).ToDeltaDiscoveryResponse()
+			resp := model.ResourceList(rs).ToDeltaDiscoveryResponse()
 			actual, err := util_proto.ToYAML(resp)
 
 			Expect(err).ToNot(HaveOccurred())
