+++ pkg/xds/generator/proxy_template_test.go
@@ -145,7 +145,7 @@ var _ = Describe("TemplateProxyGenerator", func() {
 				Expect(err).ToNot(HaveOccurred())
 
 				// then
-				resp := generator.ResourceList(rs).ToDeltaDiscoveryResponse()
+				resp := model.ResourceList(rs).ToDeltaDiscoveryResponse()
 				actual, err := util_proto.ToYAML(resp)
 				Expect(err).ToNot(HaveOccurred())
